# Log sucessful gRPC request as INFO