--- a/src/tonic/logging.rs
+++ b/src/tonic/logging.rs
@@ -58,7 +58,7 @@ where
                     if let Some(grpc_status) = grpc_status {
                         match grpc_status.code() {
                             Code::Ok => {
-                                log::trace!("gRPC {} Ok {:.6}", method_name, elapsed_sec);
+                                log::info!("gRPC {} Ok {:.6}", method_name, elapsed_sec);
                             }
                             Code::Cancelled => {
                                 // cluster mode generates a large amount of `stream error received: stream no longer needed`
@@ -95,7 +95,7 @@ where
                             ),
                         };
                     } else {
-                        log::trace!("gRPC {} Ok {:.6}", method_name, elapsed_sec);
+                        log::info!("gRPC {} Ok {:.6}", method_name, elapsed_sec);
                     }
                     Ok(response_tonic)
                 }