This PR logs successful gRPC requests as INFO.

The main motivation is to have consistency with the REST API which is already doing that and to enable users of cloud to see their gRPC logs.

![grpc-info](https://github.com/qdrant/qdrant/assets/606963/ac5bd942-af25-4736-9b3e-404a3c8d33e2)


One negative aspect is that the terminal will be much more active during benchmarking after this.