@ffuugoo 2023-09-18
Wait, so we already have logging, just on the `trace` level? I'd then just ask cloud to enable `qdrant::tonic::logging=trace`, instead of switching it to the `info`.