@timvisee 2023-09-18
Just wanted to run some numbers on this because we had some concerns about performance.

I did a very rudimentary benchmark; 10 runs of 1 million bfb upsertions.

- Without logging (by default):
  ```bash
  $ ./target/perf/qdrant
  $ hyperfine -w0 -r10 "bfb -n 1000000 --skip-wait-index --indexing-threshold 10000000000000"

  Benchmark 1: bfb -n 1000000 --skip-wait-index --indexing-threshold 10000000000000
    Time (mean ± σ):     15.146 s ±  0.190 s    [User: 2.786 s, System: 0.846 s]
    Range (min … max):   14.830 s … 15.480 s    10 runs
  ```
- With logging (by default):
  ```bash
  $ QDRANT__LOG_LEVEL=qdrant::tonic::logging=warn,info ./target/perf/qdrant
  $ hyperfine -w0 -r10 "bfb -n 1000000 --skip-wait-index --indexing-threshold 10000000000000"

  Benchmark 1: bfb -n 1000000 --skip-wait-index --indexing-threshold 10000000000000
    Time (mean ± σ):     15.765 s ±  0.208 s    [User: 2.363 s, System: 0.719 s]
    Range (min … max):   15.386 s … 16.175 s    10 runs
  ```

So in this case it increases the upsertion time by about 4% if you trust the numbers.