@agourlay 2023-09-18
Closing this one because it could potentially generate a very large amount of logs.

It is unclear if we actually want/need this and we access logs is generally provided by a proxy in front of a deployment.

Let's keep the status quo for now and re-evaluate later if needed.