@agourlay 2023-09-18
I am fine with making possible for users to have more control over the logging as long as it is not on by default all the time :)